test: split Python client coverage
